Prevent import cycle between inMemoryCache.ts and readFromStore.ts (#8850)
